refactor: merge foxy-query-builder and foxy-query-builder-rule into a single custom element
