+++ src/elements/public/QueryBuilder/index.ts
@@ -3,9 +3,7 @@ import '@polymer/iron-icons';
 import '@polymer/iron-icon';
 
 import { QueryBuilder } from './QueryBuilder';
-import { QueryBuilderRule } from './QueryBuilderRule';
 
-customElements.define('foxy-query-builder-rule', QueryBuilderRule);
 customElements.define('foxy-query-builder', QueryBuilder);
 
-export { QueryBuilder, QueryBuilderRule };
+export { QueryBuilder };
